Brings all database functions under one cli command.
