+++ examples/ticks-example.yaml
